backend/nlp_engine/requirements.txt
tensorflow==2.13.0
protobuf==3.20.3
torch==2.0.1
pandas
numpy==1.23.3
pymongo==4.4.0
flask
tensorflow_text
python-dotenv
mysql-connector-python
